Try to fix the profile image issuse
I was trying to use Jquery, but it doesn't work fine.
The scripts were left in the index.html.
Right now, I set the height of the <p> to be a fixed value and overflow to be auto.
But I don't like this.
Will try to have a better solution.